Integración con chats
