+++ RT4Sentinel/requirements.txt
@@ -1,6 +1,8 @@
 fastapi
 fastapi[standard]
 python-dotenv
+pydantic
+pytz
 requests
 sqlalchemy
 uvicorn
