--- RT4Sentinel/v1/api/chats/utils.py
@@ -1,5 +1,0 @@
-def send_message(body:dict):
-    """
-    This function sends a message to a Telegram bot.
-    """
-    print(f"Message sent to Telegram: {body}")
