+++ RT4Sentinel/v1/chats/telegram/TeleUtils/constants.py
