--- RT4ServicesProject/APIControler/api/v1/api.py
@@ -1,6 +1,0 @@
-from fastapi import APIRouter
-from .routes import routes
-
-api_router = APIRouter()
-api_router.include_router(routes.router, prefix="/v1")
-
