--- RT4ServicesProject/APIControler/api/v1/services/configs/json/groups.json
@@ -1,6 +1,0 @@
-{
-    "GrupoModelo":{
-        "actives": [],
-        "name": "Cerveceria grupo modelo"
-    }
-}
