--- RT4ServicesProject/TeleBot/docs/README.md
@@ -1,6 +1,0 @@
-# Bot Telegram
-
-Para iniciar un bot es necesario realizar pre-configuraciones para obtener el Token de Telegram que permita inciar el bot, para ello:
-1. Se debe bucar @botfather en Telegram
-1. Enviar /start -> /newbot
-1. Definir el nombre del bot, para este proyecto se elegió el nombre de RT4NetBot
