--- RT4ServicesProject/TeleBot/requirements.txt
@@ -1,6 +1,0 @@
-python-dotenv
-pydantic
-pytz
-requests
-sqlalchemy
-uvicorn
