--- RT4ServicesProject/WABot/docs/twilio.txt
@@ -1,1 +1,0 @@
-174PBT9RZ7H13LB219VL4HDH
