--- RT4ServicesProject/WABot/requirements.txt
@@ -1,1 +1,0 @@
-sqlalchemy
